build - update to latest version of redux
