+++ src/constants.ts
@@ -1,0 +1,4 @@
+export enum Statuses {
+  ERROR,
+  SUCCESS,
+}
